bundle: Add parcel + npm scripts
Use parcel to begin with as it's more plug'n play than rollup. Create a
"dist/jsxapi.min.js" and a "dist/jsxapi.js" UMD distribution bundles
exposing the library globally as "jsxapi". Using parcel seems to "just
work".

Issues with Rollup include not properly bundling and distributing node
builtins. In particular we use EventEmitter from "events", which is
possible to have resolved using rollup-plugin-node-builtins. However it
does not seem to be a complete implementation of the node.js original
module:

https://github.com/calvinmetcalf/rollup-plugin-node-builtins/pull/48

We use ".off()", so that's a show-stopper.

Fix #44