schema: Add XAPI ctor to connect factory function
`connect` is a convenience function to simplify connecting to devices.
It should be generic and support extending with e.g. typed XAPI
implementations. This change adds a XAPI constructor parameter to
`connectOverload`, defaulting to XAPI.

Using a curried version here to simplify usage for clients. This closes
around initBackend and default opts, so users won't have to handle
those.