schema: Add Config arrays
Configuration arrays are of fixed size and may be sparse (e.g.
Audio.Input.HDMI[2,3]). This means that it's not really an array, but a
sub-tree so this is also how it's best to represent this in the types.
The API calling convention still must use "array notation" `[]` in order
to properly add numbers to the path.

The following is correct:

xapi.Config.Audio.Input.HDMI[2].Mode.get();

while:

xapi.Config.Audio.Input.HDMI.2.Mode.get();

is a syntax error.