schema: Reuse Config/Status trees with generics
In order to support getting and listening to not just leaf nodes, but
also sub-trees, it's convenient that the Config and Status trees just
reflect the node types. This allows the type system to rewrite the type
of the tree for use in different situations.

E.g. `xapi.Config` and `xapi.Status` allows fetching data, but when the
data comes back the result type does not allow access to the same getter
method. There is no `.get()` and `.on()` methods on the *result* of a
fetch, but the tree structure is the same.