xapi: Add eventEmitter handler asynchronously
Do not add the event handler to the eventEmitter until the registration
is successful. Some backends (WebSocket) send multiple feedback payloads
if there are multiple expressions which match a given feedback path.
Event handlers should only be fired once, and so we wish to filter this
out based on the registration Id.

With this change there's a slight possibility (e.g. HTTP
backends/transports) that feedback is received out of order before the
registration response is received. This feedback will be lost since it
won't be emitted to any handlers.